Convert to typescript and move to composeWebhook
